src/test/java/org/example/ImpressoraTest.java
package org.example;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class ImpressoraTest {

    @Test
    void deveVerificarEquipamento() {
        Impressora impressora = new Impressora();
        impressora.verificarEquipamento();
        assertTrue(true);
    }

    @Test
    void deveRealizarVerificacoes() {
        Impressora impressora = new Impressora();
        impressora.realizarVerificacoes();
        assertTrue(true);
    }

    @Test
    void deveExecutarManutencao() {
        Impressora impressora = new Impressora();
        impressora.executarManutencao();
        assertTrue(true);
    }

    @Test
    void deveRegistrarManutencao() {
        Impressora impressora = new Impressora();
        impressora.registrarManutencao();
        assertTrue(true);
    }

    @Test
    void deveRealizarManutencaoCompleta() {
        Impressora impressora = new Impressora();
        impressora.realizarManutencao();
        assertTrue(true);
    }
}